making version 0.21.2 of tokenizers locked in project deps.
due to a new version of tokenizers==0.21.4 released hours ago, the installation fails because the pypi package didn't supply *any* of the platform binaries.

while the issue might be resolved soon, it's unknown when, and will probably will happen in the next release of tokenizers as well.

the issue might be not trivial for most people.